requirements.txt
flask
tensorflow
numpy
pillow
seaborn
